# Faster initializer tsort