### Motivation / Background

I noticed while profiling my app that initializer sorting is a bit slow. The app is very large, and has many initializers, so this makes sense. However, I think it could be better:

<img width="417" alt="Screenshot 2024-11-12 at 6 36 23 PM" src="https://github.com/user-attachments/assets/d965f035-179a-4c08-84d5-d9dee57a9c5e">

This Pull Request has been created because iterating initializers for each node child is expensive. Instead, we can use a hash to speed up tsorting children to iterate less often. This can save a lot of time on larger apps:

Before:
```
Warming up --------------------------------------
  tsort initializers    24.000 i/100ms
Calculating -------------------------------------
  tsort initializers    244.212 (± 2.5%) i/s    (4.09 ms/i) -      1.224k in   5.015190s
```

After:
```
Warming up --------------------------------------
  tsort initializers   179.000 i/100ms
Calculating -------------------------------------
  tsort initializers      1.791k (± 1.1%) i/s  (558.27 μs/i) -      9.129k in   5.097031s
```

### Detail

Benchmark:
```rb
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  gem "rails", path: "..."
  gem "benchmark-ips"
end

require "rails/all"

class MyApp < Rails::Application
  config.eager_load = false
end


Benchmark.ips do |x|
  x.report("tsort initializers") do
    Rails.application.initializers.tsort_each { }
  end
end
```

This Pull Request changes `Rails::Initializable::Collection` to sort when an initializer is added, so each node child can be found quicker. This patch saves ~400ms on my app.

### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Unrelated changes should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [x] Tests are added or updated if you fix a bug or add a feature.
* [ ] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.

cc @kmcphillips